==> pipeline_script/1.IOAggregation.ipynb <==
import pandas as pd
import networkx as nx
import numpy as np
import os
from sklearn.mixture import GaussianMixture 
from sklearn.decomposition import PCA
import datetime
import operator
import warnings
warnings.filterwarnings('ignore')

RecordWritingPath = '../result/'
TransportationDataPath = '../transportation/'
EventDataPath = '../events/'
comboPath = '../combo/'
# dataFile = TransportationDataPath+city+'EdgeYearwiseAggregated.csv'

# generate data for website only
for city in ['Taipei','NewYork']:
    print(city)
    data = pd.read_csv(TransportationDataPath+city+'EdgeDatewiseAggregated.csv')
    startId = data.groupby(['start_id','date']).sum()
    endId = data.groupby(['end_id','date']).sum()
    startPivot = pd.pivot_table(startId, values='amount', index=['date'],
                        columns=['start_id'], aggfunc=sum, fill_value=0)
    startPivot.columns = [str(x)+'start' for x in startPivot.columns]
    endPivot = pd.pivot_table(endId, values='amount', index=['date'],
                        columns=['end_id'], aggfunc=sum, fill_value=0)
    endPivot.columns = [str(x)+'end' for x in endPivot.columns]
    pd.concat([startPivot,endPivot],axis=1).to_csv(TransportationDataPath+city+'InOutFlow.csv')
print('done')

# generate data for website only
for city in ['Taipei','NewYork']:
    print(city)
    data = pd.read_csv(TransportationDataPath+city+'EdgeDatewiseAggregated.csv')
    data.groupby(['date']).sum().to_csv(TransportationDataPath+'output/'+city+'Timeseries.csv')
print('done')
