==> pipeline_script/3_1.GMM_Component_Selection.ipynb <==
import pandas as pd
import numpy as np
import os
from sklearn.mixture import GaussianMixture 
from sklearn.metrics import silhouette_score
import warnings
import matplotlib.pyplot as plt
from sklearn.decomposition import PCA
warnings.filterwarnings('ignore')

RecordWritingPath = '../result/'
TransportationDataPath = '../transportation/'
EventDataPath = '../events/'


# highest sihouette in timeseries and different n comp
standardize = 'Normalize'
city = 'Taipei'

data = pd.read_csv(TransportationDataPath+city+'EdgeDatewiseAggregated.csv')
data = pd.DataFrame(data.groupby(['date'])['amount'].sum())
matrix = data.values
silhouette = []
for ncomp in range(2,6):
    gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
    labels = gm.fit_predict(matrix)
    score = silhouette_score(matrix,labels)
    silhouette += [score]
    print(ncomp,score)
                
                

# lowest bic in timeseries and different n comp
standardize = 'Normalize'
city = 'Taipei'

data = pd.read_csv('/Users/hemingyi/Documents/capstone/post/transportation/'+city+'EdgeDatewiseAggregated.csv')
data = pd.DataFrame(data.groupby(['date'])['amount'].sum())
matrix = data.values
bic = []
for ncomp in range(2,6):
    gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
    labels = gm.fit_predict(matrix)
    score = gm.bic(matrix)
    bic += [score]
    print(ncomp,score)
                
                

# lowest aic in timeseries and different n comp
standardize = 'Normalize'
city = 'Taipei'

data = pd.read_csv('/Users/hemingyi/Documents/capstone/post/transportation/'+city+'EdgeDatewiseAggregated.csv')
data = pd.DataFrame(data.groupby(['date'])['amount'].sum())
matrix = data.values
bic = []
for ncomp in range(2,6):
    gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
    labels = gm.fit_predict(matrix)
    score = gm.aic(matrix)
    bic += [score]
    print(ncomp,score)

# highest sihouette in whole network and different n comp
standardize = 'Normalize'
city = 'Taipei'

data = pd.read_csv('/Users/hemingyi/Documents/capstone/post/transportation/'+city+'EdgeDatewiseAggregated.csv')
data['edge'] = data['start_id'] + '-' + data['end_id']
data = data.drop(['start_id','end_id'], axis=1)
data = data.pivot(index='date',columns='edge').reset_index()
data.fillna(0,inplace=True)
matrix = data.drop(['date'], axis=1).values
pca = PCA(n_components=16)    
matrix=pca.fit_transform(matrix)
silhouette = []
for ncomp in range(2,6):
    gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
    labels = gm.fit_predict(matrix)
    score = silhouette_score(matrix,labels)
    silhouette += [score]
    print(ncomp,score)
                
                

for ncomp in range(2,6):
    gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
    labels = gm.fit_predict(matrix)
    score = gm.bic(matrix)
    print(ncomp,score)

# highest sihouette in IO and different n comp
standardize = 'Normalize'
city = 'Taipei'

data = pd.read_csv(TransportationDataPath+'IO/OriginSize/Normalize/'+city+'IONormalize.csv')

matrix = data.drop(['date'], axis=1).values
silhouette = []
for ncomp in range(2,6):
    gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
    labels = gm.fit_predict(matrix)
    score = gm.bic(matrix)
    silhouette += [score]
    print(ncomp,score)
                
                

standardize = 'Normalize'
city = 'Taipei'

data = pd.read_csv(TransportationDataPath+'IO/OriginSize/Normalize/'+city+'IONormalize.csv')

matrix = data.drop(['date'], axis=1).values
silhouette = []
for ncomp in range(2,6):
    gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
    labels = gm.fit_predict(matrix)
    score = gm.aic(matrix)
    silhouette += [score]
    print(ncomp,score)
                

standardize = 'Normalize'
city = 'NewYork'

data = pd.read_csv('/Users/hemingyi/Documents/capstone/post/transportation/'+city+'EdgeDatewiseAggregated.csv')
data = pd.DataFrame(data.groupby(['date'])['amount'].sum())
matrix = data.values
silhouette = []
for ncomp in range(2,6):
    gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
    labels = gm.fit_predict(matrix)
    score = silhouette_score(matrix,labels)
    silhouette += [score]
    print(ncomp,score)

# highest sihouette in different methods and different n comp
standardize = 'Normalize'
city = 'Taipei'
for aggregation in ['Comm','IO']:
    for dimension in ['OriginSize','PCA','AE']:
        data = pd.read_csv(TransportationDataPath+'%s/%s/%s/%s%s%s.csv'%(aggregation, dimension, standardize, city, aggregation, standardize))
        matrix = data.drop(['date'], axis=1).values
        print(aggregation, dimension)
        silhouette = []
        for ncomp in range(2,6):
            gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
            labels = gm.fit_predict(matrix)
            score = silhouette_score(matrix,labels)
            silhouette += [score]
        bestcomp = np.array(silhouette).argmax() + 2
        print(bestcomp)
                
                

# same n component, average sihouette for different methods
standardize = 'Normalize'
city = 'Taipei'
for ncomp in range(2,6):
    silhouette = []
    print(ncomp)
    for aggregation in ['Comm','IO']:
        for dimension in ['OriginSize','PCA','AE']:
            data = pd.read_csv(TransportationDataPath+'%s/%s/%s/%s%s%s.csv'%(aggregation, dimension, standardize, city, aggregation, standardize))
            matrix = data.drop(['date'], axis=1).values
#             print(aggregation, dimension)
            gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
            labels = gm.fit_predict(matrix)
            score = silhouette_score(matrix,labels)
            silhouette += [score]
    meansilhouette = np.array(silhouette).mean()
    print(meansilhouette)

                

## aic

standardize = 'Normalize'
city = 'Taipei'
for aggregation in ['Comm','IO']:
    for dimension in ['OriginSize','PCA','AE']:
        data = pd.read_csv(TransportationDataPath+'%s/%s/%s/%s%s%s.csv'%(aggregation, dimension, standardize, city, aggregation, standardize))
        matrix = data.drop(['date'], axis=1).values
        print(aggregation, dimension)
        aic = []
        for ncomp in range(2,6):
            gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
            gm.fit(matrix)
            score = gm.aic(matrix)
            aic += [score]
        bestcomp = np.array(aic).argmin() + 2
        print(bestcomp)
                
                

# New York

## silhouette

standardize = 'Normalize'
city = 'NewYork'
for aggregation in ['Comm','IO']:
    for dimension in ['OriginSize','PCA','AE']:
        data = pd.read_csv(TransportationDataPath+'%s/%s/%s/%s%s%s.csv'%(aggregation, dimension, standardize, city, aggregation, standardize))
        matrix = data.drop(['date'], axis=1).values
        print(aggregation, dimension)
        silhouette = []
        for ncomp in range(2,6):
            gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
            labels = gm.fit_predict(matrix)
            score = silhouette_score(matrix,labels)
            silhouette += [score]
        bestcomp = np.array(silhouette).argmax() + 2
        print(bestcomp)
                
                

## bic

standardize = 'Normalize'
city = 'NewYork'
for aggregation in ['Comm','IO']:
    for dimension in ['OriginSize','PCA','AE']:
        data = pd.read_csv(TransportationDataPath+'%s/%s/%s/%s%s%s.csv'%(aggregation, dimension, standardize, city, aggregation, standardize))
        matrix = data.drop(['date'], axis=1).values
        print(aggregation, dimension)
        bic = []
        for ncomp in range(2,6):
            gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
            gm.fit(matrix)
            score = gm.bic(matrix)
            bic += [score]
        bestcomp = np.array(bic).argmin() + 2
        print(bestcomp)
                
    

## aic

standardize = 'Normalize'
city = 'NewYork'
for aggregation in ['Comm','IO']:
    for dimension in ['OriginSize','PCA','AE']:
        data = pd.read_csv(TransportationDataPath+'%s/%s/%s/%s%s%s.csv'%(aggregation, dimension, standardize, city, aggregation, standardize))
        matrix = data.drop(['date'], axis=1).values
        print(aggregation, dimension)
        aic = []
        for ncomp in range(2,6):
            gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
            gm.fit(matrix)
            score = gm.aic(matrix)
            aic += [score]
        bestcomp = np.array(aic).argmin() + 2
        print(bestcomp)
                
                

# lowest bic in timeseries and different n comp
standardize = 'Normalize'
city = 'NewYork'

data = pd.read_csv('/Users/hemingyi/Documents/capstone/post/transportation/'+city+'EdgeDatewiseAggregated.csv')
data = pd.DataFrame(data.groupby(['date'])['amount'].sum())
matrix = data.values
bic = []
for ncomp in range(2,6):
    gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
    labels = gm.fit_predict(matrix)
    score = gm.bic(matrix)
    bic += [score]
    print(ncomp,score)
                
                

# highest sihouette in timeseries and different n comp
standardize = 'Normalize'
city = 'NewYork'

data = pd.read_csv('/Users/hemingyi/Documents/capstone/post/transportation/'+city+'EdgeDatewiseAggregated.csv')
data = pd.DataFrame(data.groupby(['date'])['amount'].sum())
matrix = data.values
silhouette = []
for ncomp in range(2,6):
    gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
    labels = gm.fit_predict(matrix)
    score = silhouette_score(matrix,labels)
    silhouette += [score]
    print(ncomp,score)
                
                

# highest sihouette in whole network and different n comp
standardize = 'Normalize'
city = 'NewYork'

data = pd.read_csv('/Users/hemingyi/Documents/capstone/post/transportation/'+city+'EdgeDatewiseAggregated.csv')
data['edge'] = data['start_id'].astype('str') + '-' + data['end_id'].astype('str')
data = data.drop(['start_id','end_id'], axis=1)
data = data.pivot(index='date',columns='edge').reset_index()
data.fillna(0,inplace=True)
matrix = data.drop(['date'], axis=1).values
pca = PCA(n_components=16)    
matrix=pca.fit_transform(matrix)
silhouette = []
for ncomp in range(2,6):
    gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
    labels = gm.fit_predict(matrix)
    score = silhouette_score(matrix,labels)
    silhouette += [score]
    print(ncomp,score)
                
                

for ncomp in range(2,6):
    gm=GaussianMixture(n_components=ncomp,n_init=100,max_iter=1000,random_state=0)
    labels = gm.fit_predict(matrix)
    score = gm.bic(matrix)
    print(ncomp,score)
